new features added

+++ routes.py
@@ -3,6 +3,7 @@ from flask import render_template,request, redirect, url_for, flash,session
 from models import db,User,ParkingLot,ParkingSpot,Reservation, Vehicle
 from werkzeug.security import generate_password_hash, check_password_hash
 from functools import wraps
+from datetime import datetime
     
 @app.route('/login')
 def login():
@@ -95,7 +96,7 @@ def admin_required(f):
         user= User.query.get(session['user_id'])
         if not user.is_admin:
             flash('You do not have permission to access this page')
-            return redirect(url_for('index'))
+            return redirect(url_for('user'))
         return f(*args, **kwargs)
     return decorated_function
 
@@ -114,12 +115,12 @@ def admin():
 @app.route('/lot/add')
 @admin_required
 def add_lot():
-    return render_template('addlot.html')
+    return render_template('lot/addlot.html')
 
 @app.route('/lot/add' ,methods=['POST'])
 @admin_required
 def add_lot_post():
-    
+    # creating lot
     name= request.form.get('name')
     address = request.form.get('address')
     pincode = request.form.get('pincode')
@@ -132,19 +133,22 @@ def add_lot_post():
     new_lot=ParkingLot(name=name,address=address,pincode=pincode,price_per_hour=price,total_spots=totalspots)
     db.session.add(new_lot)
     db.session.commit()
-    return redirect(url_for('admin'))
+    totalspots = int(totalspots)  # Convert to integer 
+    # creating parking spots
+    for i in range(1, totalspots + 1):
+        spot_number = i
+        new_spot = ParkingSpot(lot_id=new_lot.id, spot_number=spot_number, is_occupied=False)
+        db.session.add(new_spot)
 
-
-@app.route('/lot/<int:id>/')
-@admin_required
-def view_lot(id):
-    return "view lot"    
+    db.session.commit()
+    flash("Parking lot and spots created successfully!")
+    return redirect(url_for('admin')) 
 
 @app.route('/lot/<int:id>/edit')
 @admin_required
 def edit_lot(id):
     parkinglot = ParkingLot.query.get(id)
-    return render_template('editlot.html', parkinglot=parkinglot)
+    return render_template('lot/editlot.html', parkinglot=parkinglot)
 
 @app.route('/lot/<int:id>/edit', methods=['POST'])
 @admin_required
@@ -161,18 +165,41 @@ def edit_lot_post(id):
         return redirect(url_for('edit_lot', id=id))
     
     db.session.commit()
+
+    updated_spots = int(request.form.get('totalspots'))
+    current_spots = ParkingSpot.query.filter_by(lot_id=id).count()
+    if updated_spots > current_spots:
+        # Add new parking spots
+        for i in range(current_spots + 1, updated_spots + 1):
+            new_spot = ParkingSpot(lot_id=id, spot_number=i, is_occupied=False)
+            db.session.add(new_spot)
+
+    elif updated_spots < current_spots:
+        spots_to_delete = (ParkingSpot.query.filter_by(lot_id=id).order_by(ParkingSpot.spot_number.desc()).limit(current_spots - updated_spots).all())
+        for spot in spots_to_delete:
+            db.session.delete(spot)
+
+    db.session.commit()
+    flash('Parking lot updated successfully')
     return redirect(url_for('admin'))
 
 @app.route('/lot/<int:id>/delete')
 @admin_required
 def delete_lot(id):
     parkinglot = ParkingLot.query.get(id)
-    return render_template('deletelot.html', parkinglot=parkinglot)
+    return render_template('lot/deletelot.html', parkinglot=parkinglot)
+
 
 @app.route('/lot/<int:id>/delete', methods=['POST'])
 @admin_required
 def delete_lot_post(id):
     parkinglot = ParkingLot.query.get(id)
+    occupied_spot=next((s.id for s in parkinglot.spots if s.is_occupied), None)
+    if occupied_spot:
+        flash("You can't delete the lot as spot is occupied in this lot.")
+        return redirect(url_for('admin'))
+
+    # Delete parking lot if there is no spot occupied
     db.session.delete(parkinglot)
     db.session.commit()
     return redirect(url_for('admin'))
@@ -182,7 +209,7 @@ def delete_lot_post(id):
 @admin_required
 def userdata():
     users = User.query.all()
-    return render_template('userdata.html', users=users)
+    return render_template('user/userdata.html', users=users)
 
 
 @app.route('/')
@@ -197,9 +224,7 @@ def index():
 @app.route('/profile')
 @auth_required
 def profile():
-    user = {
-        'Name': session['name']
-    }
+    user = User.query.get(session['user_id'])
     return render_template("profile.html", user=user)
     
 
@@ -207,7 +232,6 @@ def profile():
 @auth_required
 def edit_profile():
     user= User.query.get(session['user_id'])
-    admin= User.query.filter_by(is_admin=True).first()
     return render_template('editprofile.html',user=user)
 
 @app.route('/profile/edit', methods=['POST'])
@@ -250,31 +274,149 @@ def edit_profile_post():
 @auth_required
 def user():
     parkinglots = ParkingLot.query.all()
-    return render_template('user.html',parkinglots=parkinglots)
+    return render_template('user/user.html',parkinglots=parkinglots)
     
 @app.route('/lot/<int:id>/book')
 @auth_required
 def book_lot(id):
     parkinglot = ParkingLot.query.get(id)
+    vehicles = Vehicle.query.filter_by(user_id=session['user_id']).all()
+    spot = next((s.id for s in parkinglot.spots if not s.is_occupied), None)
+    if not spot:
+        flash('No vacant spots available in this lot.')
+        return redirect(url_for('user'))
+    if not vehicles:
+        flash('No vehicles found. Please add a vehicle first.')
+        return redirect(url_for('vehicle'))
     if not parkinglot:
         flash('Parking lot not found')
         return redirect(url_for('user'))
+    return render_template('lot/booklot.html', parkinglot=parkinglot,vehicles=vehicles, spot=spot)
+
+
+@app.route('/lot/<int:id>/book', methods=['POST'])
+@auth_required
+def book_lot_post(id):
+    lot_id = request.form.get('lot_id')
+    vehicle_id = request.form.get('vehicleid')
+    spot_id= int(request.form.get('spotid'))
+
+    # Validate vehicle 
+    vehicle = Vehicle.query.filter_by(id=vehicle_id, user_id=session['user_id']).first()
+    if not vehicle:
+        flash(f'Vehicle {vehicle_id} not found or not owned by the user.')
+        return redirect(url_for('vehicle')) 
     
-    # Check if the user has already booked a spot in this lot
-    # existing_reservation = Reservation.query.filter_by(user_id=session['user_id'], parkinglot_id=id).first()
-    # if existing_reservation:
-    #     flash('You have already booked a spot in this parking lot.')
-    #     return redirect(url_for('user'))
+    # timestamp
+    # timestamp=now.strftime("%H:%M:%S")
     
-    return render_template('booklot.html', parkinglot=parkinglot)
-    
+    # Check if the user has already booked for given vehicle
+    existing_reservation = Reservation.query.filter_by(user_id=session['user_id'], lot_id=id,vehicle_id=vehicle.id).first()
+    if existing_reservation:
+        flash(f'You have already booked a spot for vehicle : {vehicle.vehicle_number} in this parking lot.')
+        return redirect(url_for('parking_history'))
+
+    # Create reservation
+    reservation = Reservation(
+        user_id=session['user_id'],
+        lot_id=lot_id,
+        spot_id=spot_id,
+        vehicle_id=vehicle.id,
+        parking_timestamp=datetime.now(),
+        leaving_timestamp=None,
+        total_cost=None
+    )
+    spot=ParkingSpot.query.filter_by(id=spot_id).first()
+    # Mark the spot as occupied
+    spot.is_occupied = True
+    db.session.add(reservation)
+    db.session.commit()
+
+    flash(f'Spot successfully reserved ')
+    return redirect(url_for('user'))
+
+@app.route('/reserved/<int:id>/view')
+@admin_required
+def view_reserve(id):
+    reservation=Reservation.query.get(id)
+    leaving_timestamp = datetime.now()
+    parking_timestamp= reservation.parking_timestamp
+
+    duration = leaving_timestamp - parking_timestamp
+    hrs = duration.total_seconds() / 3600
+    est_cost = round(hrs * reservation.parkinglot.price_per_hour, 2)
+    return render_template('spot/occupiedspot.html',reservation=reservation,est_cost=est_cost)
+
+
 @app.route('/parking/history')
 @auth_required
 def parking_history():
-    parkinglots= ParkingLot.query.all()
-    return render_template('parkinghistory.html',parkinglots=parkinglots)
+    reservations=Reservation.query.filter_by(user_id=session['user_id'],leaving_timestamp=None).all()
+    return render_template('user/parkinghistory.html',reservations=reservations)
 
-@app.route('/release')
+@app.route('/spot/<int:id>/release')
 @auth_required
-def release():
-    return render_template('release.html')
+def release(id):
+    reservation=Reservation.query.get(id)
+    leaving_timestamp = datetime.now()
+    parking_timestamp= reservation.parking_timestamp
+
+    duration = leaving_timestamp - parking_timestamp
+    hrs = duration.total_seconds() / 3600
+    cost = round(hrs * reservation.parkinglot.price_per_hour, 2)
+
+    return render_template('spot/release.html',reservation=reservation,l_timestamp=leaving_timestamp,cost=cost)
+
+@app.route('/spot/<int:id>/release', methods=['POST'])
+@auth_required
+def release_post(id):
+    reservation = Reservation.query.get(id)
+    l_str=request.form.get('l_timestamp') 
+    # converting str to datetime
+    reservation.leaving_timestamp = datetime.fromisoformat(l_str)
+    reservation.total_cost=request.form.get('cost')
+    spot=ParkingSpot.query.filter_by(id=reservation.spot_id).first()
+    # Delete parking lot
+    spot.is_occupied=False
+    db.session.commit()
+    flash('Spot released successfully')
+    return redirect(url_for('parking_history'))
+
+
+@app.route('/vehicle')
+@auth_required
+def vehicle():
+    vehicles=Vehicle.query.filter_by(user_id=session['user_id']).all()
+    return render_template('user/vehicle.html',vehicles=vehicles)
+
+
+@app.route('/vehicle/add' )
+@auth_required
+def add_vehicle():
+    return render_template('user/addvehicle.html') 
+
+@app.route('/vehicle/add', methods=['POST'])
+@auth_required  
+def add_vehicle_post():
+    user_id = session['user_id']
+    vehicle_number = request.form.get('vehicleno')
+    db.session.add(Vehicle(user_id=user_id, vehicle_number=vehicle_number))
+    db.session.commit()
+    flash('Vehicle added successfully')
+    return redirect(url_for('vehicle'))
+
+# searching lot
+@app.route('/search')
+@admin_required
+def search_lot():
+    return render_template('lot/searchlot.html')
+
+
+#-----------------------------  spot--------------------------------------
+
+
+@app.route('/admin/deletespot/<int:id>')
+@admin_required
+def delete_spot(id):
+    spot=ParkingSpot.query.get(id)
+    return render_template('spot/deletespot.html',spot=spot)
